chore: clean up legacy configuration and documentation files
- Remove deprecated GitHub workflows, issue templates, and CI/CD configurations
- Delete obsolete Husky git hooks and pre-commit scripts
- Remove completed Kiro specification directories and task files
- Delete legacy test scripts and temporary development files
- Remove Filament v4 migration documentation and release-please configs
- Clean up VSCode settings, Makefile, and performance/security audit reports
- Consolidate security and performance improvements into core application files
- Add new security middleware and optimization services for production readiness
- Implement comprehensive boot error profiling and rate limiting enhancements
- Add secure query builder and password handling traits for enhanced security
- Expand test coverage with security audit and performance integration tests
- Update language files and CORS configuration for improved localization
- Streamline repository structure by removing completed migration artifacts and temporary files

+++ app/Jobs/StorePerformanceMetricsJob.php
@@ -34,7 +34,7 @@ class StorePerformanceMetricsJob implements ShouldQueue
         private readonly string $routeName,
         private readonly float $ttfb,
         private readonly int $queryCount,
-        private readonly int $peakMemoryMb,
+        private readonly float $peakMemoryMb,
         private readonly string $environment,
         private readonly ?string $userAgent = null
     ) {
